chore: move acn protocol tests
