+++ packages/hashes.csv
@@ -36,8 +36,8 @@ open_aea/contracts/scaffold,bafybeiba4pdufnxve5hdkla42lixnxwfkyg7s2bvuy6zqbxjsxt
 open_aea/protocols/scaffold,bafybeihhyjuqvp5zqtikkjrwtikg4fr4tsub43n777huxtpeu77g75poze
 open_aea/protocols/signing,bafybeihjlpgjm5vkg6kxm66a5k2r357dadnsjfpddta2mzd2bshstv6vdi
 open_aea/skills/scaffold,bafybeih3h7yia76rrbittnoeotpraiuho5rdqsnksuvncinaa3akgmnbaa
-valory/connections/p2p_libp2p,bafybeigxbnvgurf6byb646snqchqlrpxfd4h3rxt7bvg5zlg3yjcq7tzma
-valory/connections/p2p_libp2p_client,bafybeif7plocy4y6bvhtb3bob3onsyt6lycyjzcdfk5ze72qa2xwqb4yji
-valory/connections/p2p_libp2p_mailbox,bafybeie6mbdudc6iryro6pxjxsbkx2xnmexvthb7rwkum2pt6tv4n26aey
-valory/protocols/acn,bafybeif7dt6cwbixufftupr2srfuzg2w2vgevtqdxwq3z25yab3ptpkyqq
+valory/connections/p2p_libp2p,bafybeiebn2fe6sankudbesknoz4oqa5ir6fzyzlzmcadlserwxx637gwau
+valory/connections/p2p_libp2p_client,bafybeibjsohzs4gyjothhiskcrmcnsxtlkf5tq2aipbzh7vyjkq2ky4bsm
+valory/connections/p2p_libp2p_mailbox,bafybeigyxfnmpvkmmvxkbkni3bjd7q4q27pm7zkw5tczr54tqriuozms7m
+valory/protocols/acn,bafybeig5eegg2pc3clffga6fhyta2jt7so3nt4eqlvkzaiqmia26luzxlm
 valory/protocols/tendermint,bafybeidbdvqgrr2lbstyqbyb24kxhen6xnv65e3zqg7xj6fwrogrzms2ty
