+++ packages/valory/connections/p2p_libp2p/connection.yaml
@@ -58,7 +58,7 @@ fingerprint_ignore_patterns: []
 build_entrypoint: check_dependencies.py
 connections: []
 protocols:
-- valory/acn:1.1.0:bafybeif7dt6cwbixufftupr2srfuzg2w2vgevtqdxwq3z25yab3ptpkyqq
+- valory/acn:1.1.0:bafybeig5eegg2pc3clffga6fhyta2jt7so3nt4eqlvkzaiqmia26luzxlm
 class_name: P2PLibp2pConnection
 config:
   delegate_uri: 127.0.0.1:11000
