+++ packages/valory/connections/p2p_libp2p_mailbox/connection.yaml
@@ -14,7 +14,7 @@ fingerprint:
 fingerprint_ignore_patterns: []
 connections: []
 protocols:
-- valory/acn:1.1.0:bafybeif7dt6cwbixufftupr2srfuzg2w2vgevtqdxwq3z25yab3ptpkyqq
+- valory/acn:1.1.0:bafybeig5eegg2pc3clffga6fhyta2jt7so3nt4eqlvkzaiqmia26luzxlm
 class_name: P2PLibp2pMailboxConnection
 config:
   connect_retries: 3
