+++ packages/valory/protocols/acn/protocol.yaml
@@ -15,6 +15,8 @@ fingerprint:
   dialogues.py: bafybeicgcqcnztuwcwvwzzbh5jjmc5gw56yzk65b33gxj7fu7caoay6bd4
   message.py: bafybeiggjefapn3d5s623fazxrl7von5yqel4eczrie5yvwk4s7gbumt44
   serialization.py: bafybeiaz3ykvi6nyyx6cdry7fepy76l55jpduhl6vazhmrgt7trejmksam
+  tests/__init__.py: bafybeidteufp2npjd77ekcftk5e4gbaquq3gike5nxtk5xfmnusls56keu
+  tests/test_acn.py: bafybeia4usywpj5gnp5qgdne4xojwwnx2jowvjx3xo53v6bl262fjzkhsq
 fingerprint_ignore_patterns: []
 dependencies:
   protobuf: {}
